main: split leds_test_http.c
