+++ main/http_routes.h
@@ -42,6 +42,7 @@ int leds_api_get(struct http_request *request, struct http_response *response, v
 int leds_api_get_status(struct http_request *request, struct http_response *response, void *ctx);
 int leds_api_post(struct http_request *request, struct http_response *response, void *ctx);
 
+/* leds_test_http.c */
 int leds_api_test_get(struct http_request *request, struct http_response *response, void *ctx);
 int leds_api_test_post(struct http_request *request, struct http_response *response, void *ctx);
