+++ main/leds_http.c
@@ -223,41 +223,3 @@ int leds_api_get_status(struct http_request *request, struct http_response *resp
 
   return 0;
 }
-
-/* GET /api/leds/test */
-static int leds_api_write_test_array(struct json_writer *w)
-{
-  int err;
-
-  for (const struct config_enum *e = leds_test_mode_enum; e->name; e++) {
-    if ((err = JSON_WRITE_OBJECT(w,
-      JSON_WRITE_MEMBER_STRING(w, "mode", e->name)
-    ))) {
-      return err;
-    }
-  }
-
-  return 0;
-}
-
-static int leds_api_write_test(struct json_writer *w, void *ctx)
-{
-  return JSON_WRITE_ARRAY(w, leds_api_write_test_array(w));
-}
-
-int leds_api_test_get(struct http_request *request, struct http_response *response, void *ctx)
-{
-  int err;
-
-  if ((err = http_request_headers(request, NULL))) {
-    LOG_WARN("http_request_headers");
-    return err;
-  }
-
-  if ((err = write_http_response_json(response, leds_api_write_test, NULL))) {
-    LOG_WARN("write_http_response_json -> leds_api_write_test");
-    return err;
-  }
-
-  return 0;
-}
